refactor(cluster): split distributor service state support tests
